requirements-lambda.txt
# Core Orca SDK (REQUIRED)
orcapt-sdk>=1.0.3

# AWS SDK (REQUIRED for Lambda)
boto3>=1.34.0

# FastAPI to Lambda adapter (REQUIRED for HTTP endpoints)
mangum>=0.17.0

# OpenAI SDK (REQUIRED for AI functionality)
openai>=1.12.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.25.0

